Changed planet position to double because of precision issues

+++ planet.h
@@ -22,7 +22,7 @@ public:
   glm::mat4 view_mat;
   glm::vec3 view_pos;
   glm::vec3 light_pos;
-  glm::vec3 view_center;
+  glm::dvec3 view_center;
   Shader &planet_shader, &atmos_shader, &sun_shader, &ring_shader;
   Renderable &planet_obj, &atmos_obj, &ring_obj;
 };
@@ -33,7 +33,7 @@ public:
   Orbit();
   void setParameters(const std::string &parent_body, double ecc, double sma, double inc, double lan, double arg, double m0);
   void computePosition(double epoch); /// Computes the position from current epoch and parent body
-  const glm::vec3 &getPosition() const;
+  const glm::dvec3 &getPosition() const;
   void setParentFromName(std::deque<Planet> &planets);
   bool isUpdated();
   void reset();
@@ -43,7 +43,7 @@ private:
   bool updated;
   std::string parent_body;
 
-  glm::vec3 position;
+  glm::dvec3 position;
   Planet *parent;
 };
 
@@ -93,7 +93,7 @@ public:
   void load();
   void unload();
   void update(double epoch);
-  void render(const glm::vec3 &pos, const RenderContext &rc, const Ring &rings, const Atmosphere &atmos);
+  void render(const glm::dvec3 &pos, const RenderContext &rc, const Ring &rings, const Atmosphere &atmos);
 
   void print() const;
 
@@ -138,7 +138,7 @@ public:
   void print() const;
 
   const std::string &getName() const;
-  const glm::vec3 &getPosition() const;
+  const glm::dvec3 &getPosition() const;
 
   Orbit &getOrbit();
   Body &getBody();
